src/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
openai>=1.3.5
beautifulsoup4==4.12.2
requests==2.31.0
pydantic>=2.0.0
httpx<1.0.0
